refactor(votes): move SenateVoteDetail XML parsing into the factory method
The previous from_senate_xml was a one-line delegate to
senate_xml.parse_vote_detail — exactly the cargo-cult pattern ADR 0018
Rule 2 was meant to prevent. Move the orchestration AND the
SenateVoteDetail-specific helpers (_resolve_subject, _iter_positions,
_build_amendment_id, _build_bill_id, _SENATE_THRESHOLD_MAP,
_SENATE_BILL_TYPE_MAP, _parse_senate_date, _xml_text_to_int) into
models/votes.py so the factory method owns the parsing logic in one
place, collocated with the model schema.

Also remove the dead @field_validator _coerce_chamber from
SenateVoteDetail. chamber is hard-coded "senate" in the constructor
(this class only parses senate.gov XML), so the validator never fired;
beyond being dead code, it violated ADR 0018 Rule 3 which bans
semantic-shim field_validators on wire-shape models.

senate_xml.py is now down to the HTTP client + the two parsers that
return non-model types (parse_vote_menu → list[int], parse_senate_roster
→ dict[str, str]) + general XML utilities. The previously-circular
models ↔ senate_xml import is gone — senate_xml no longer needs
anything from concord.models.

Tests for the moved private helpers (TestParseSenateDate,
TestBuildSenateIds) move to tests/test_models_votes.py alongside the
helpers they exercise; the parse_vote_detail end-to-end tests in
test_senate_xml.py become SenateVoteDetail.from_senate_xml integration
tests against the same XML fixtures, validating the orchestration.

All 638 tests pass; mypy strict clean; ruff clean.
